Initial commit
If you're building this and you're not me, the glide.yaml isn't
set up to do that back here in the past just yet (I need to be
able to build it without having pushed my CPU to github). Grab
virt6502 yourself or change the yaml to get around that.